Add AbstractIdentifiable class file
